fix: synchronize reservation/client/agence lists, remove dead code, unify monetary types to double, add VoyageNational.terminer()

+++ src/main/java/TransportTerrestre/Depenses.java
@@ -17,7 +17,7 @@ public class Depenses {
     }
 
     private LocalDate date;
-    private int montant;
+    private double montant;
     private String description;
     @ToString.Exclude @EqualsAndHashCode.Exclude
     private Vehicule vehicule;
